Spliting header file
